v2.0.389.0: Decimal ratings, chosen by sweeping the stars
Issue #82, asked for by three people: 7.5 rather than being forced to pick 7 or
8. Alorski would have settled for half steps; this does tenths.

Server: UserRating.Rating, UserRatingDetail.Rating and RatingStats.UserRating
widen from int to double, and the write endpoint takes a double in 0.1-10.
Files written by older versions hold whole numbers, which deserialize into a
double unchanged, so there is no migration. Rounding to one decimal happens once
inside SetRatingAsync rather than at the call sites, so nothing - including a
hand-made API call - can put 7.30000000000000004 in the file.

Everything that bucketed on an exact rating now rounds to a whole star first:
the two profile distributions, the dashboard distribution and browse-by-rating.
Left alone they would have counted almost nothing once decimals existed, and the
distribution arrays are indexed by that value.

Client: the star row is a continuous control. Moving over it previews the value
under the pointer and fills the stars to match, using the same
background-clip:text gradient the half-star state used, generalised from a fixed
50% to any fraction. Pointer events cover mouse, touch and pen, so a swipe on a
phone works the same way; touch-action keeps the page from scrolling mid-swipe.

Three things live testing caught. Pressing in the gap between two stars did
nothing, because the handler demanded a hit on the glyph - it now accepts the
whole row. The value chip resized as its text changed and, with the row
centre-justified, shoved the stars sideways under the pointer, which made the
reading jump while sweeping; it has a fixed width now. And the release position
is read on pointerup rather than reusing the last move.

The number is shown while pointing at the row and stays visible once the item is
rated. Visibility is driven by :hover in CSS rather than a JS flag so no code
path can leave it stuck on. Committing still requires a press, so passing over
the stars never changes a saved rating.

+++ Models/RatingStats.cs
@@ -30,7 +30,7 @@ namespace Jellyfin.Plugin.Ratings.Models
         /// Gets or sets the user's rating (if applicable).
         /// </summary>
         [JsonPropertyName("userRating")]
-        public int? UserRating { get; set; }
+        public double? UserRating { get; set; }
 
         /// <summary>
         /// Gets or sets the rating distribution (count for each rating value 1-10).
